new geo map for welcome page
