+++ client/src/pages/welcome.js
@@ -2,64 +2,24 @@
 // import getCurrentPosition from '../components/currentPosition';
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { Button } from "@progress/kendo-react-buttons";
+import { Map, InfoWindow, Marker, GoogleApiWrapper } from "google-maps-react";
 
-let map, infoWindow;
-
-function initMap() {
-  map = new google.maps.Map(document.getElementById("map"), {
-    center: { lat: -34.397, lng: 150.644 },
-    zoom: 6,
-  });
-  infoWindow = new google.maps.InfoWindow();
-
-  // Try HTML5 geolocation.
-  if (navigator.geolocation) {
-    navigator.geolocation.getCurrentPosition(
-      (position) => {
-        const pos = {
-          lat: position.coords.latitude,
-          lng: position.coords.longitude,
-        };
-        infoWindow.setPosition(pos);
-        infoWindow.setContent("Location found.");
-        infoWindow.open(map);
-        map.setCenter(pos);
-      },
-      () => {
-        handleLocationError(true, infoWindow, map.getCenter());
-      }
-    );
-  } else {
-    // Browser doesn't support Geolocation
-    handleLocationError(false, infoWindow, map.getCenter());
-  }
-}
-
-function handleLocationError(browserHasGeolocation, infoWindow, pos) {
-  infoWindow.setPosition(pos);
-  infoWindow.setContent(
-    browserHasGeolocation
-      ? "Error: The Geolocation service failed."
-      : "Error: Your browser doesn't support geolocation."
-  );
-  infoWindow.open(map);
-}
-
-class ButtonContainer extends React.Component {
+export class MapContainer extends Component {
   render() {
     return (
-      <div className="row example-wrapper">
-        <div className="col-xs-12 col-sm-6 example-col">
-          <p>
-            <Button primary={true}>Profile</Button>;
-            <Button primary={true}>Map</Button>;
-            <Button primary={true}>Messages</Button>;
-          </p>
-        </div>
-      </div>
+      <Map google={this.props.google} zoom={14}>
+        <Marker onClick={this.onMarkerClick} name={"Current location"} />
+
+        <InfoWindow onClose={this.onInfoWindowClose}>
+          <div>
+            <h1>{this.state.selectedPlace.name}</h1>
+          </div>
+        </InfoWindow>
+      </Map>
     );
   }
 }
 
-export default ButtonContainer;
+export default GoogleApiWrapper({
+  apiKey: YOUR_GOOGLE_API_KEY_GOES_HERE,
+})(MapContainer);
